+++ bc/search/tests/test_fis_pages.py
@@ -25,3 +25,26 @@ class FISPagesIncludedTest(TestCase):
         self.assertIn(
             Page.objects.get(pk=hit_page.pk), response.context["search_results"]
         )
+
+
+class FISPagesExcludedTest(TestCase):
+    def setUp(self):
+        self.fis_homepage = fixtures.SubsiteHomePageFactory()
+
+    def test_fis_page_not_in_search_results(self):
+        image = ImageFactory()
+        hit_page = self.fis_homepage.add_child(
+            instance=fixtures.SubsiteHomePageFactory.build(
+                title="Buckinghamshire Local Government Pension Scheme",
+                listing_summary="Buckinghamshire Local Government Pension Scheme",
+                hero_image=image,
+                banner_image=image,
+                is_pensions_site=True,
+            )
+        )
+
+        response = self.client.get(reverse("search") + "?query=pension")
+
+        self.assertNotIn(
+            Page.objects.get(pk=hit_page.pk), response.context["search_results"]
+        )